+++ CHANGES.md
@@ -14,6 +14,7 @@
  - [ ] Give a try to Kivy in another package (userflow first)
  - [ ] Apply 10^log10 to model for Kinetic in order to reduce errors
  - [ ] Find out why some conversion ratio are not shown in steady mode
+ - [ ] Solve issue when KineticFit Solver p0 is too far away from real give shape error 
 
 ## v0.1.14